examples/main_agent_controller/requirements.txt
openai>=1.40.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
jsonschema>=4.23.0
httpx>=0.27.0
